Fairly significant API overhaul
TIPManifest is now package-protected and is only used internally.  Client
code should use TIPPackageFactory to open or create a package and then
interact with the fields through a package object.  This simplifies 
use of the API and also makes it possible to do more cleanup in the future.

As a combination of this commit and the last, it's now easier to create 
response packages based on existing request packages, and also to use the
new TIPTaskType interface to create standard or custom packages.